Comment out tests that use testbed lease envs
We've had trouble with communication between the telemetry_receiver and testbed lease envs. It seems to be related to certs of our receiver (app name telemetry-acceptance-loader) not being trusted by the testbed lease envs. We think if we were able to obtain a subdomain of *.pivotal.io with a load balancer in front (like we have for the production loader) the issue would be solved. However, at the moment, there aren't any new subdomains being issued for that domain. We'll pick this back up once we are able to obtain a new domain for a our loader.